feat: reuse over duplicate

+++ crates/cdk/src/publish.rs
@@ -66,11 +66,7 @@ impl PublishCmd {
             package_info.package_path().to_string_lossy()
         );
 
-        if hubdir.exists() {
-            tracing::warn!("Removing directory at {:?}", hubdir);
-            remove_dir_all(&hubdir)?;
-        }
-
+        self.cleanup(&package_info)?;
         build_connector(&package_info, BuildOpts::with_release(opt.release.as_str()))?;
         init_package_template(&package_info, &self.readme)?;
         check_package_meta_visiblity(&package_info)?;
@@ -84,7 +80,6 @@ impl PublishCmd {
                     .unwrap_or_else(|| hubdir.join(hubutil::HUB_PACKAGE_META));
                 let pkgdata = package_assemble(pkgmetapath, &opt.target, &access)?;
                 package_push(self, &pkgdata, &access)?;
-                remove_dir_all(&hubdir)?;
             }
 
             // --pack only
@@ -104,8 +99,23 @@ impl PublishCmd {
                     .clone()
                     .ok_or_else(|| anyhow::anyhow!("package file required for push"))?;
                 package_push(self, pkgfile, &access)?;
-                remove_dir_all(&hubdir)?;
             }
+        }
+
+        if !self.pack {
+            self.cleanup(&package_info)?;
+        }
+
+        Ok(())
+    }
+
+    pub fn cleanup(&self, package_info: &PackageInfo) -> Result<()> {
+        let hubdir = package_info.package_relative_path(DEF_HUB_INIT_DIR);
+
+        if hubdir.exists() {
+            // Delete the `.hub` directory if already exists
+            tracing::warn!("Removing directory at {:?}", hubdir);
+            remove_dir_all(&hubdir)?;
         }
 
         Ok(())
